[Core] Add ManagerFactory
Closes #507. `Session` should be migrated to a host-convenience library.
It is simply a state management helper designed to aid monolithic widget
based C++ apps manage state. It is certainly not an interaction pattern
required by the API, and complicates a number of simple host
implementation scenarios.

Hence migrate business logic from `Session` into a simple
`ManagerFactory`. The `Session` class (and its usages) will be removed
as part of future work.

The `ManagerFactory` is available in both C++ and Python, allowing C++
hosts to execute the salient logic of `Session`.

The `identifiers` member function provides a list of identifiers, as
supplied by the provided `ManagerInterfaceFactoryInterface` instance.

The `availableManagers` method returns a map of manager identifier to
`ManagerDetail` structs, each containing the `identifier`, `displayName`
and `info`. These details are extracted from instantiated
`ManagerInterface` objects, on the assumption that they are cheap to
construct.

The `createManager` function finally allows construction of a `Manager`,
making use of the `HostInterface` held in the `ManagerFactory` plus an
`identifier` argument used to instantiate the `ManagerInterface`, which
are both fed into the `Manager` constructor.

The `createManagerForInterface` `static` member function provides a
convenient single entry point for direct construction of a `Manager`.
This allows bypassing the construction of a `ManagerFactory` instance
when the identifier is known ahead of time. Bindings to other languages
may prefer to skip the `ManagerFactory` class as a whole and simply bind
to this one static member function. Pybind does not allow binding two
overloads when one of them is a `static` member function, hence to
maintain parity between C++ and Python we could not also name this
funciton `createManager`.

Since the `Manager` and `HostSession` classes in Python currently have
much more functionality than in C++, we must ensure that a
`ManagerFactory` used from within Python will return a Python `Manager`
instance.

So add a Python subclass that reimplements the `createManager` and
`createManagerForInterface` methods. This is a temporary measure until
the `Manager`/`HostSession` C++ migration is complete.

Ensure signature parity with `createManagerForInterface` between C++ and
Python, i.e. take a `LoggerInterface` instance, even though it is not
currently required by the C++ `HostSession` constructor.

Signed-off-by: David Feltell <david.feltell@foundry.com>