[Core] Remove `Session` class
Closes #510.

Session is simply a state management helper designed to aid monolithic
widget based C++ apps manage state. It is certainly not an interaction
pattern required by the API, and complicates a number of simple host
implementation scenarios.

So remove all usages of `Session` from code and documentation in favour
of the new, simpler, `ManagerFactory`, which is now responsible for
querying available manager implementations and creation of `Manager`
wrappers.

Signed-off-by: David Feltell <david.feltell@foundry.com>