[Core] Add and use `PyRetainingSharedPtr`
Python instances can be passed to and held by a C++ instance as a C++
base class pointer. The Python instance can then be garbage collected,
leaving the base class pointer "dangling", resulting in "RuntimeError
Tried to call pure virtual function" from the pybind 'trampoline'.

So add a `PyRetainingSharedPtr` class with a custom `type_caster` that
ensures the Python object is kept alive whilst the `PySharedPtr` (and
`shared_ptr`s constructed from it) is alive.

Add a `RetainPyArgs` compile-time helper class that is given a (member)
function pointer and a list of `shared_ptr` types and will decorate the
function such that any arguments to the function that match the given
`shared_ptr` types are converted to instead be `PyRetainingSharedPtr`s.

Add `RetainCommonPyArgs` as an alias to `RetainPyArgs`, with pre-filled
template parameters. The template parameters are those types whose
Python-derived instance(s) must be kept alive.

Add a new test target and module to check this new behaviour. Hence move
all Python CMake related to testing under this new target.

For each Python bound method/property that accepts or returns a
`shared_ptr` from Python to C++, where the Python instance inherits from
the C++ type, use `PyRetainingSharedPtr`. In particular, where
appropriate:
- Trampolines returning a Python instance to C++ now return a
  `PyRetainedSharedPtr` via the `PYBIND11_OVERRIDE` macro.
- Constructors bound using the pybind `init<...>` helper has its
  template arguments switched to `PyRetainingSharedPtr`.
- Constructors bound using a `make` factory now use the
  `RetainCommonPyArgs` helper.
- Methods that take a Python instance now use the `RetainCommonPyArgs`
  helper.
- Member variables that can be set from Python are now properties with
  a mutator function that takes a `PyRetainingSharedPtr`.

Add a section to the CODING_STANDARDS doc detailing usage of the
`PyRetainingSharedPtr`.

Signed-off-by: David Feltell <david.feltell@foundry.com>