[Core] Add createPythonPluginSystemManagerImplementationFactory
Closes #508.

The convenience method is to allow a C++ host to instantiate the Python
plugin system without needing access to a compatible version of Pybind.

Add a new `openassetio-python-bridge` target, creating a separate
`openassetio-python` C++ library, which depends only on
`openassetio-core` and uses pybind to import the Python library. Hence
assumes Python is already initialised in the wider application.

Signed-off-by: David Feltell <david.feltell@foundry.com>