[Core] Invert log severity constant order
Reverses the order of logging severity constants, such that a logically
higher severity now has a higher associated numeric value.

This is now in line with pretty much every other logging library in the
world. Lol.

We didn't feel the need to include padding between the values (such as
with python's `logging` module), as:
 - It breaks the lookup names by index trick.
 - This isn't a general purpose logging library, so introducting
   new logging levels in some specific context makes less sense.

Closes #516.

Signed-off-by: Tom Cowland <tom@foundry.com>