[Core] Migrate `ConsoleLogger` to C++
This migrates the simple console logger to C++ from Python.

A few minor tweaks to output formatting, but more notably, all messages
now go to std::err to avoid polluting program output.

This also drops the facaded output workarounds from the Python version
until we're in a position to see if they're needed from C++, and find
an alternative approach if they are.

Part of #542.

Signed-off-by: Tom Cowland <tom@foundry.com>