[Core] Add `EntityReference` type
Adds the EntityReference value semantic type to be used to represent
a manager validated string such that it can be passed to entity related
API methods.

We opted for header-only implementation for now as the implementation
is currently trivial.

We use a concrete class with constructor (vs struct with public members)
to make it conceptually opaque to support future extensions to include
manager identifier validation.

Part of #549.

Signed-off-by: Tom Cowland <tom@foundry.com>