[Core] Make `isEntityReferenceString` singular
When we batch-first the API, we made `isEntityReferenceString` a batch
method. In hindsight, this complicates common use cases, without
saving much in terms of overhead because it is always an in-process
mechanism.

This call is always meant to be a light-weight check to see if some
string originated from a specific manager to be used as an entity
reference, not that it is a definitively valid reference. Thinking of
this in terms of an URL-based entity reference, simply checking the URL
scheme is that picked by the manager would be sufficient.

With this in mind, the flow control overhead of batching up strings
for testing is no less in a batch form of this method (as they first
need packing, then the results unpacking and processing). In the
singular form, you can do this as part of the processing step without
the need for allocating temporary single element arrays numerous times.

Because the per-call overhead for this is low in comparison to entity
data related calls, then wrapping this up in a batch version for
functional programming situations should not suffer the overhead that
such an approach would incur for methods such as resolve, that talk to
some distant back-end service.

Part of #549.

Signed-off-by: Tom Cowland <tom@foundry.com>