[Core] Migrate `isEntityReferenceString` to C++
Part of #549. In order for a C++ host to use an entity reference string
as a parameter to manager API methods, the host is required to validate
that it is relevant to the target manager.

Currently, this is expected to be an explicit validation by the host
application, but in upcoming work this will also be performed by entity
reference factory functions.

So as a prerequisite step toward C++ migration of manager API methods
that use entity references, migrate `isEntityReferenceString` to C++, so
that entity reference strings can be validated before use.

Signed-off-by: Tom Cowland <tom@foundry.com>