[Core] Add `Manager.createEntityReference`
Part of #549. Added the `createEntityReference` method to C++ and
Python, which should always be used to create `EntityReference` objects.

The method will validate that the provided entity reference string is
relevant to the manager and throw a `std::domain_error` if it isn't
(exception type to be migrated to a custom type in future work).

Signed-off-by: David Feltell <david.feltell@foundry.com>