[Core] Add `Manager.createEntityReferenceIfValid`
Part of #549. Added the `createEntityReferenceIfValid` method to C++ and
Python, which can be used as an alternative to `createEntityReference`
for creating `EntityReference` objects.

The method returns an `optional`, which only contains a value if the
the provided entity reference string is relevant to the manager.

The `optional` specialisation type is not aliased with a typedef, as
`std::optional` is well known and conveys the behaviour without
requiring further documentation.

Signed-off-by: David Feltell <david.feltell@foundry.com>