[Core] Add BatchElementError
Closes #557.

Our batch-first design means that errors that affect only a single
element of the batch should not cause the whole batch to fail. Instead,
such errors are communicated via an error callback, with a unique call
per failed batch element.

We need some way for the manager to communicate the reason for a failure
to the error callback. In order for this to be extensible we need to
bundle the error details into a class. At minimum these details will
include an error code and message, giving the same level of information
as a standard exception.

So add a simple `BatchElementError` aggregate class, initially
containing just an error code and string message. The error code is
a strongly typed enumeration, currently only supporting a fallback
code of `kUnknown`.

To prevent ambiguity of these error codes with the C API's exception
error codes, and to allow reuse of the error codes between C++ and C,
add a new C-compatible header the core library, with `#define`s
computing the new error codes as offset from the C API's exception error
codes.

Signed-off-by: David Feltell <david.feltell@foundry.com>