[Core] Migrate `resolve` to a callback-based API
Part of #533. The new design from #483 (DR009) for batch-first methods
involves adding success and error callbacks, to be called per element
(entity reference, in this case) with the results.

So switch over to this new scheme for `resolve`.

This is due to be migrated from Python to C++ imminently, so not much
effort has been put into documenting the callback signature, since that
will be handled automatically by Doxygen for the C++ implementation.

Signed-off-by: David Feltell <david.feltell@foundry.com>