[Core] Add kInvalidEntityReference batch error
Part of #533. Support invalid entity references being passed to
`resolve`, which are pre-validated with `isEntityReferenceString`, but
nevertheless malformed for the particular entity.

The docs and usage surrounding this was previously limited and
ambiguous. So add an `apiComplianceSuite` check to make it concrete,
with BAL extended to make use of this error code.

Use `InvalidBALReference` in `bal.py` instead of
`InvalidEntityReference`. This allows us to remove the OpenAssetIO
dependency of the BAL 'application' code.

Refactor BAL's `resolve` implementation to make it clearer which calls
can raise an exception.

Signed-off-by: David Feltell <david.feltell@foundry.com>