[Core] Remove `Str` typedef
Having solicited broader consensus around the use of typedefs for
simple types, the conclusion was that aliasing `std::string` to `Str`
adds more conceptual overhead than is justifiable by the potential
future benefits of an abstraction.

Principally, casual reading of the code now calls a spade a `std::spade`
which means the reader already understands the semantics of the type.

We have no intention to allow customization of this type at build time,
so the main merits in future maintenance are hard to justify if it
makes learning the code based harder for an integrator.

As there is no compiler checking of the miss-use of these types of
aliases, we had also ended up with situations where the wrong one was
accidentally used - potentially creating further ambiguity as to the
correct approach.

Closes #527.

Signed-off-by: Tom Cowland <tom@foundry.com>