[Core] Remove enum utility aliases
Part of #568.

The `EnumIdx` alias was a convenience to allow indexing into an
`EnumNames` array without the need for `static_cast`. However, the enums
are due to be switched to `enum class` strongly typed scoped enums,
which require a `static_cast` to extract the underlying value
regardless, so the enforcement of `EnumIdx` as the enum type is
basically useless (it _could_ be a micro-optimisation, but is premature
at this point).

The `EnumNames` alias was used for associating particular enum values
with `std::string_view`s wrapping string literals containing the enum's
human-readable name. This alias is not very useful in practice. The
compiler should optimise string literals when used[^1], so there's
no need to pre-wrap in a `std::string_view`. If we do want a
`std::string_view`, it's trivial to wrap on-demand, at zero cost.

Removing the use of these aliases significantly simplifies the code
without incurring any real penalty.

[^1] In practice, concatenation of a string literal to a `std::string`
in GCC<=9.5 isn't clever, but GCC>=10, Clang and VC++ are clever enough
to avoid the computation of the length of a string literal (according to
my interpretation of the disassembly).

Signed-off-by: David Feltell <david.feltell@foundry.com>