[Tests] Use `Identifier` typedef alias in test
It's unfortunate that there appears to be no linter check available to
ensure the appropriate alias is used.

We can argue that its fine to use a different "type" to store the return
value of a function, as long as it is implicitly convertible (which is
obviously the case for two aliases to the same type). However, one of
the reasons for using aliases is as semantic documentation, which we
lose if we go mixing and matching aliases.

Signed-off-by: David Feltell <david.feltell@foundry.com>