[Core] Fix C++ hosts calling Python manager's `resolve`
Part of #582. Whilst writing an experimental C++ end-to-end test, it was
discovered that passing a C++ `std::function` to `Manager.resolve`
fails with a `TypeError`, since pybind does not know how to convert the
`std::function` to a Python callable. We must include
`pybind11/functional.h` to enable that support.

Signed-off-by: David Feltell <david.feltell@foundry.com>