[Tests] Add checks for source of `Manager` implementation
Inspired by #582, where we inadvertently neglected to delete the Python
implementation when migrating to C++.

Attempts to ensure that the method is implemented in the expected
language during the Python -> C++ transition. The approach is slightly
tenuous, but seem to work ok.

Methods like `getsource` could also be used, but ultimately, they raise
a `TypeError` if the supplied object isn't a function... so this seems
simpler to invoke/manager.

Signed-off-by: Tom Cowland <tom@foundry.com>