[Core] Remove default `ManagerInterface.preflight` implementation
The intent of the default implementation was to simplify writing a
manager plugin when there are no placeholder/reservation steps
required. However, it is not possible for a default implementation
to assert that the supplied references are considered writable for
publishing. This would result in `preflight` succeeding, but `register`
then failing with a kEntityAccesssError - _after_
expensive work had been done. Ensuring `preflight` errors in the
same scenarios is essential in avoiding potentially huge amounts of
redundant work.

Part of #587.

Signed-off-by: Tom Cowland <tom@foundry.com>