[Core] Add `ManagerInterface._createEntityReference`
In #549, we added a 'strongly typed' entity reference workflow. This
required hosts to ensure that strings are validated through the manager
before being used as an entity reference. This ensures that once initial
checks have been completed, manager implementations no longer need
to worry about re-checking method inputs to see if they are one of their
references.

We plan on adding assorted validation checks in the API middleware that
will help avoid coding errors when multiple managers are in use.

This raises the question of how should a manager implementation create
these strongly typed objects - where it doesn't need the validation?

We wanted to avoid direct construction of `EntityReference` objects
on the stack by code so that we can evolve the manager validation aspect
with minimal code churn at all the construction sites.

We previously debated an object oriented design
(`EntityReference(string, manager)` vs the current function
+ data approach (`Manager.createEntityReference(string)`). We picked the
latter as it matched the approach we have with `Context` and is more
portable to other languages.

This seems to have lead to a simple solution to this question - where by,
we add an alternate (protected) factory method to the ManagerInterface
that bypasses the usual checks performed when the strings originate from
outside the manager's implementation.

This should then be used by all manager implementation code that needs
to create an entity reference to return to the host.

Signed-off-by: Tom Cowland <tom@foundry.com>