[Core] Rename malformed reference fixtures in `apiComplianceSuite`
Renames `a_malformed_reference` to `an_invalid_reference`, and
`a_malformed_entity_refrence` to `a_malformed_reference`.

The original `a_malformed_refrence` was used as the opposite of
`a_valid_reference` when verifying the behavior of
`isEntityReferenceString`.  Making it the inverse better covers what it
should be.

We also had 'a_malformed_entity_reference' which was a valid reference
with some subtle error in its actual value (eg: incorrect query param).
Renamed this to just `a_malformed_refrence`. This should allow us to
more globally use "invalid" to mean "not one of your references", and
"malformed" to mean "it is one of yours, but isn't quite right".

Signed-off-by: Tom Cowland <tom@foundry.com>