[Core] Add `defaultManagerForInterface` to `ManagerFactory`
Adds a convenience to `ManagerFactory` that allows a "default" manager
to be created from a shared settings file. This allows deployments to
create a single source of truth that is easily referenced by all API
hosts, without a need for their own settings persistence mechanism.

This simplifies the centralised management of OpenAssetIO config.

The OPENASSETIO_DEFAULT_CONFIG environment variable can be
set to point to a simple TOML file with the following structure:

```toml
[manager]
identifier = "some.identifier"

[manager.settings]
some_key = "a value"
```

The manager settings section is optional.

We opted to use TOML as the config file is intended to be easily
authored/edited by hand, and the data structures are always simple.

In the future this could easily be extended such that the env var can
list a number of fallback config files if required, or to allow the
direct specification of the TOML path in the call.

There was some contemplation between `OPENASSETIO_DEFAULT_MANAGER_CONFIG`
Vs `OPENASSETIO_DEFAULT_CONFIG`. In the end, the latter was selected as
it has scope to be used for other aspects of the API (eg: controlling
debug/auditing functionality that will be added before “v1”).

There is still a “default manager config” constant on `ManagerFactory`
though, incase we choose to de-couple the above. With the thinking that
“this is the var that controls the `defaultManagerForInferface` method,
rather than “what is a common, shared config var”. The former works in
both this situation, and a potential later one where this mechanism
may be controlled independently.

This commit does not change LICENSE or NOTICE as if I’ve understood
https://infra.apache.org/licensing-howto.html#bundled-vs-non-bundled
correctly, we shouldn’t need it until we start creating binary
distributions.

Closes #494.

Signed-off-by: Tom Cowland <tom@foundry.com>