[Core] Forward `ManagerFactory.py` methods to C++ and re-wrap interface
ManagerFactory was fairly trivial when first authored. We will soon be
adding more functionality to it. To avoid duplicate business logic,
this switches exposing a Manager's interface - allowing us to re-use
the construction/initialization logic from C++, and re-wrap the
ManagerInterface instance in a Python Manager to expose the full
Python API to Python hosts.

We can remove all this when Manager.py is removed.

Part of #494.

Signed-off-by: Tom Cowland <tom@foundry.com>