[Core] Fix `RetainPyArgs` for const ref args
The `RetainPyArgs` helper was using strict type matching when comparing
function arguments against its `PtrsToPyRetain` list. This meant, for
example, that a `const shared_ptr<SomeType>&` argument would not match
because of the const reference qualifiers.

The decorated function returned from `RetainPyArgs::decorator` would
then be missing the expected `PyRetainingSharedPtr<SomeType>`
argument(s), so the custom pybind type caster wouldn't trigger on that
argument to extend the lifetime of the associated Python instance.

So ensure the argument type is `std::decay`ed, to remove any qualifiers,
before checking for a match.

Signed-off-by: David Feltell <david.feltell@foundry.com>