[CI] Support updating conan cache
Whilst implementing `ccache` in CI for #613, it became apparent that
the conan cache should use the same strategy, in order for us to be able
to add new conan packages and not have to manually bust the cache.

Specifically, we create an entirely new cache with every run (but with a
constant prefix on the cache key). This means loading the cache fails
and instead uses the `restore-keys` fallback, which will match the most
recently used cache that matches a prefix, i.e. the previous run.

This may also fix or otherwise be relevant for #612, where we
occasionally found the cache was evicted for some unknown reason.

Signed-off-by: David Feltell <david.feltell@foundry.com>