[Examples] Add `simpleResolver` CLI host example
Adds a very simple CLI host, that resolves the supplied entity reference
for the given trait set. The resulting data is written to stdout as JSON.

This was a sprint demo, but since the addition of the
OPENASSETIO_DEFAULT_CONFIG mechanism in #608, its generally a lot more
useful to the wider world.

Signed-off-by: Tom Cowland <tom@foundry.com>