[Build] Don't use pip to install Python component
Part of #629. Avoid using `pip` to install the core pure Python
component of OpenAssetIO.

Instead, used CMake's `install(DIRECTORY` to bundle the Python sources
in a proto-environment, upon which a Python virtual environment can
optionally be overlaid.

This has several advantages
* In future we can use `setup.py` (`pip`) to puppet CMake, without fear
  of infinite recursion.
* It gives us an intermediate bundle directory that can be used
  independently or for further packaging, where a Python environment is
  unnecessary/undesirable.
* It improves install times (CMake is clever enough to only reinstall
  changed files), useful for the TDD loop.

Hence remove now-unused `setup.py` - leaving it in is confusing and
broken, even though we aim to reintroduce it in a working form soon.

Signed-off-by: David Feltell <david.feltell@foundry.com>