[Tests] Allow codegen tests to run without installing
Part of #629. We wish to allow tests to be executed in an external
Python environment with `PYTHONPATH` augmented, rather than installing
`openassetio_codegen` as a package. However, doing so means that the
`console_scripts` (defined in `setup.py`) will not be created, breaking
the tests.

So switch command-line invocation tests from using the
`openassetio-codegen` script to using the active Python interpreter
executable. This leaves a small hole in our test coverage, which will
be handled by #651.

Signed-off-by: David Feltell <david.feltell@foundry.com>