[Build] Add setup.py that builds Python extension
Part of #630. Reintroduce `setup.py`, this time with the ability to
puppet CMake to build the Python extension module, so that direct
execution of CMake is not required. I.e. `pip install .` from the repo
root works, provided CMake's `find_package` can locate dependencies.

As noted in the `bootstrap-windows-2019.bat` CI bootstrap script,
ConanCenter's `cpython` package on Windows is incomplete unless we set
the option `cpython:shared=True`. Unfortunately, adding that option
breaks Linux builds, so add a TODO to conditionally enable the option
once we migrate to `conanfile.py`.

Signed-off-by: David Feltell <david.feltell@foundry.com>