[Build] Migrate to `conanfile.py` from `.txt`
Part of #630. In order to better support Windows builds using
ConanCenter's `cpython` package, we must enable the `shared` option.
However, this breaks Linux builds (due to executable vs. library RPATH
issues). So we need to enable this option only if using Windows. This
precludes using the simpler `conanfile.txt`, where such conditionality
is not possible.

So migrate to a basic `conanfile.py` that has the same effect as the
previous `conanfile.txt`, except with the aforementioned conditional
`cpython` option.

This also paves the way for future work:
- Automating Python wheel building, where we want to build against the
  external environment's Python, rather than a fixed conan `cpython`
  package.
- Packaging OpenAssetIO up as a Conan package, for redistribution.

Signed-off-by: David Feltell <david.feltell@foundry.com>