[Core] Drop minimum Python version to 3.7
This affords us compatibility as far back as VFX Platform 2020 (from a
Python perspective), with minimal loss of functionality. This year is
still in active use in production, so compatibility there makes adoption
easier.

This commit deliberately doesn't update the conanfile, as there are
ASAN fails for 3.7 runs (believed to be an upstream bug, fixed in later
versions).

Direct leak of 432 byte(s) in 1 object(s) allocated from:
    #0 0x7f7ac9c17808 in __interceptor_malloc ../../../../src/libsanitizer/asan/asan_malloc_linux.cc:144
    #1 0x55edd73a9c44 in _PyMem_RawWcsdup source_subfolder/Objects/obmalloc.c:592

We will add matrix testing of 3.7 via wheel builds, as long as they are
supported.

Part of #660.
Signed-off-by: Tom Cowland <tom@foundry.com>