[Build] Add Python info to OpenAssetIOConfig.cmake
On the road to creating a public Conan package for #654, for consumption
as a bundled CMake package it is useful to have defined the Python
version OpenAssetIO was built against, and the location of it's
installed Python sources.

This means that linked projects can ensure compatible Python ABIs and
extract the Python sources.

Signed-off-by: David Feltell <david.feltell@foundry.com>