[Build] Add ALIAS CMake targets to match exported
Tangentially related to #654. When OpenAssetIO is consumed as a CMake
package, the dependent project links to e.g.
`OpenAssetIO::openassetio-core`. However, when consumed as a subproject,
these targets are not available and the dependent project must link to
`openassetio-core` (i.e. without the namespace).

So for consistency in different usage scenarios, add CMake `ALIAS`
targets that match the exported, i.e. namespaced, targets.

Signed-off-by: David Feltell <david.feltell@foundry.com>