[Build] Support linking in the build tree
Tangentially related to #654.

When OpenAssetIO is used as a subproject and the consuming project wants
to run against the build tree, the libraries must be discoverable by
the linker.

As a specific example, linking to `openassetio-core-c` will then
transitively link to `openassetio-core`. The consuming project may have
its RUNPATH set to locate `openassetio-core-c`, but then when that tries
to transitively link in `openassetio-core` it can fail to locate it, if
`openassetio-core-c` does not have an appropriate RUNPATH of its own.

Note this is specific to RUNPATH behaviour. The older RPATH behaviour
passes on the search paths of the parent to the child library, so
wouldn't suffer this problem.

So relax our restriction, designed to enforce that tests always run from
the install tree, to allow CMake to auto-populate the RUNPATH/RPATH
fields of libraries in the build tree.

Signed-off-by: David Feltell <david.feltell@foundry.com>